+++ changelog.rst
@@ -13,10 +13,11 @@ Moldy Bread Change Log
 
 * Get a list of objects created or modified by a user with audit_responsibility.  See docs for more details.
 
-**New Fedora Public Methods**:
+**New Fedora Public Methods and Other Procs**:
 
 * FedoraRequest.audit_responsibility(username): Returns a list of objects where a specific username has created or modified the object at some point.
 * FedoraRequest.count_versions_of_datastreams(dsid): Returns pids with the total number of versions a specified datastream has.
+* process_versions(): Requires a sequence of tuples with pids and number of versions are returns matches based on the specified operation and version value.
 
 0.1.3 - January 10, 2019
 =================================